+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- support for monotonic ULID values. See `next_monotonic` and `next_strictly_monotonic`.
+
 ### Changed
 - updated dependencies.