+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][unreleased]
+
+## [0.6.0] - 2018-10-24
 ### Added
 - derive `Copy` for `Ulid`.
 
@@ -95,7 +97,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Everything. This was the initial release.
 
-[unreleased]: https://github.com/huxi/rusty_ulid/compare/0.5.0...HEAD
+[unreleased]: https://github.com/huxi/rusty_ulid/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/huxi/rusty_ulid/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/huxi/rusty_ulid/compare/0.4.1...0.5.0
 [0.4.1]: https://github.com/huxi/rusty_ulid/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/huxi/rusty_ulid/compare/0.3.0...0.4.0